ArgumentTypeClassNameVisitor -> SpecializeBinaryClassNameVisitor
